last-minute edits that didn't get demo'ed
you will have to run preprocess.py again for this to run!

Unfortunately we we worried about getting everything working properly so we stuck with an earlier version of the mapviz demo (even though I did another 3hrs to polish it up 😭). So anyway here are those edits and bug fixes + I added numbers to the pollution setting now because I wanted to do that too.
Also there was a draft -even working- version of a tape style scroll bar for time of day but I honestly value the simplicity of this more but I can push it to a branch -but nobody cares this is just perfectionism lol/ - Greg